Add files via upload
Cleaned up some code: Got rid of duplicate individual Subreddit data classes by making the parsable json data it's own class. I also created a method to pull the data rather than call the same code each time.
